Moved tokens to separate header
