+++ include/sparse/parsers/title_parser.hpp
@@ -1,11 +1,16 @@
 #pragma once
 
-#include<optional>
-#include<string_view>
-#include<sparse/common/types.hpp>
+#include <optional>
+#include <sparse/common/types.hpp>
+#include <string_view>
 
-namespace sparse::parsers {
-
+namespace sparse::parsers
+{
+/**
+ * @brief Parses bibliography title
+ * @param file_ stringview of the whole file
+ * @return string with title if successful, std::nullopt otherwise
+ */
 std::optional<common::title_t> parse_title(std::string_view file_);
 
 } // namespace sparse::parsers
